[refactor] split utils.py file into multiple files for better scaling and readability
